requirements-auth.txt
# Authentication System Dependencies
PyJWT==2.8.0
bcrypt==4.1.2
boto3==1.34.0
botocore==1.34.0
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.3

# Development dependencies
python-dotenv==1.0.0 